pyproject.toml
[project]
name = "mower-rover"
version = "0.1.0"
description = "CLI tooling for converting a Husqvarna Z254 zero-turn mower into an autonomous RTK-mowing robot."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Zero-Turn Mower Rover" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "structlog>=24.1",
    "pymavlink>=2.4.40",
    "pyubx2>=1.2.45",
    "shapely>=2.0",
    "pyproj>=3.6",
    "pyyaml>=6.0",
]

[project.scripts]
mower = "mower_rover.cli.laptop:app"
mower-jetson = "mower_rover.cli.jetson:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mower_rover"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "sitl: tests that require ArduPilot SITL (rover-skid frame)",
    "field: tests that require physical hardware in the field",
]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]

[tool.ruff.lint.flake8-bugbear]
# Typer relies on call expressions in defaults (Argument/Option) by design.
extend-immutable-calls = ["typer.Argument", "typer.Option"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

[[tool.mypy.overrides]]
module = ["pymavlink.*", "pyubx2.*", "shapely.*", "pyproj.*"]
ignore_missing_imports = true
